requirements.txt
torch>=2.0.0
transformers>=4.37.0
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
accelerate>=0.20.0
streamlit>=1.28.0
sentencepiece>=0.1.99
protobuf>=3.20.0
openai>=1.0.0
python-docx>=0.8.11
anthropic>=0.25.0
google-generativeai>=0.3.0